Update why-csharp.rst
I suggest changing it to read "and the .NET platform". .NET framework might confuse students as .NET Framework is a real thing and different from .NET Core which is what the course is teaching. 

C# is a programming language developed by Microsoft and used in almost all of its products might not be true.  Microsoft has a lot of products built with C, C++ and React Native.  It might be better to read "and used in many of its products."  This one was hard to fact check with recent numbers.

I don't normally here Visual Studio referred to as Visual Studio IDE, just Visual Studio, even though it is an IDE.  I understand you are trying to differentiate it from Visual Studio Code though.

FYI, in November the names Framework and Core are going away with the release of .NET 5.  Everything is getting merging into one, but really it mostly means that is something has not made it to Core from Framework like Web Forms, it never will.